Rename 'client' to 'revclient' and improve connection handling.
Updated the module and variable names to reflect 'revclient' for clarity and consistency. Enhanced connection handling by adding logging and ensuring safe resource cleanup with deferred error-checking.
